Final Refactor for best practices
File names should be snake case
