+++ src/actions/action_repo.js
@@ -1,2 +1,2 @@
 // Require All of your Application action files Here to get them registered
-import './example_store_actions/GetExampleMessageAction.js';
+import './example_store_actions/get_example_message.js';
