+++ src/actions/example_store_actions/get_example_message.js
@@ -1,6 +1,6 @@
-import Actions from '../../core/AppActions.js';
+import Actions from '../../core/app_actions.js';
 import Axios from 'axios';
-import ExampleStore from '../../stores/ExampleStore.js';
+import ExampleStore from '../../stores/example_store.js';
 import Router, { checkStatus, handleError } from '../../core/router.js';
 
 import { GET_EXAMPLE_MESSAGE } from '../../constants.js';
