+++ src/stores/example_store.js
