+++ src/test/test_store.js
